requirements.txt
dj-database-url
Django
gunicorn
psycopg2
whitenoise
pygments
djangorestframework